gateway/src/test/javascript/spec/app/entities/apibank/solicabert-conta/solicabert-conta.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { SolicabertContaService } from 'app/entities/apibank/solicabert-conta/solicabert-conta.service';
import { ISolicabertConta, SolicabertConta } from 'app/shared/model/apibank/solicabert-conta.model';

describe('Service Tests', () => {
  describe('SolicabertConta Service', () => {
    let injector: TestBed;
    let service: SolicabertContaService;
    let httpMock: HttpTestingController;
    let elemDefault: ISolicabertConta;
    let expectedResult: ISolicabertConta | ISolicabertConta[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(SolicabertContaService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new SolicabertConta(0, 'AAAAAAA', 0, 0, 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a SolicabertConta', () => {
        const returnedFromService = Object.assign(
          {
            id: 0,
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new SolicabertConta()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a SolicabertConta', () => {
        const returnedFromService = Object.assign(
          {
            nome: 'BBBBBB',
            cpf: 1,
            saldoinicial: 1,
            status: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of SolicabertConta', () => {
        const returnedFromService = Object.assign(
          {
            nome: 'BBBBBB',
            cpf: 1,
            saldoinicial: 1,
            status: 'BBBBBB',
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a SolicabertConta', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
